# src/main/resources/public/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is the Definitions Catalogue - where government defines it's digital data",
    "version" : "1.0.0",
    "title" : "Definitions Catalogue",
    "contact" : {
      "email" : "alastair.parker@ato.gov.au"
    },
    "license" : {
      "name" : "MIT",
      "url" : "https://opensource.org/licenses/MIT"
    }
  },
  "host" : "definitions.ausdx.io",
  "basePath" : "/",
  "schemes" : [ "https" ],
  "paths" : {
    "/api/definition/{domain}/{id}" : {
      "get" : {
        "tags" : [ "definitions" ],
	"operationId" : "findDefinition",
        "summary" : "Finds definitions by their id",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "domain",
          "in" : "path",
          "description" : "The shorthand name of a domain. E.g. 'sbr' for Standard Business Reporting and 'fs' for Financial Statistics",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "The id of an item. E.g. 'de55'",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Definition"
            }
          }
        }
      }
    },
    "/api/definitions" : {
      "get" : {
        "tags" : [ "definitions" ],
	"operationId" : "browseDefinitions",
        "summary" : "Browse through all the definitions",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Which result page",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "How many results per page",
          "required" : false,
          "type" : "integer"
        } ],
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Results"
            }
          }
        }
      }
    },
    "/api/search" : {
      "get" : {
        "tags" : [ "search" ],
	"operationId":"searchDefinitions",
        "summary" : "Search for definitions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "description" : "The search term",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "domain",
          "in" : "query",
          "description" : "Which domain to search within. Defaults to '', which searches all domains. Domains are specified by their shorthand notation, e.g. 'sbr' for Standard Business Reporting",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Which result page",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "How many results per page",
          "required" : false,
          "type" : "integer"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "schema" : {
              "$ref" : "#/definitions/Results"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "Definition" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "domain" : {
          "type" : "string"
        },
        "definition" : {
          "type" : "string"
        },
        "guidance" : {
          "type" : "string"
        },
        "identifier" : {
          "type" : "string",
          "format" : "uri"
        },
        "usage" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        }
      }
    },
    "Results" : {
      "type" : "object",
      "properties" : {
        "content" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Definition"
          }
        },
        "numberOfElements" : {
          "type" : "integer"
        },
        "firstPage" : {
          "type" : "boolean"
        },
        "lastPage" : {
          "type" : "boolean"
        },
        "totalPages" : {
          "type" : "integer"
        },
        "id" : {
          "$ref" : "#/definitions/Link"
        },
        "links" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Link"
          }
        }
      }
    },
    "Link" : {
      "type" : "object",
      "properties" : {
        "rel" : {
          "type" : "string"
        },
        "href" : {
          "type" : "string"
        }
      }
    }
  }
}